some codeowner reorganizing and cleanup
