+++ packages/kbn-securitysolution-exception-list-components/src/exception_item_meta_form/translations.ts
@@ -1,21 +1,22 @@
 /*
  * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
  * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
  */
 
 import { i18n } from '@kbn/i18n';
 
 export const RULE_EXCEPTION_NAME_LABEL = i18n.translate(
-  'xpack.securitySolution.rule_exceptions.itemMeta.nameLabel',
+  'exceptionList-components.exceptionItemMetaForm.nameLabel',
   {
     defaultMessage: 'Rule exception name',
   }
 );
 
 export const RULE_EXCEPTION_NAME_PLACEHOLDER = i18n.translate(
-  'xpack.securitySolution.rule_exceptions.itemMeta.namePlaceholder',
+  'exceptionList-components.exceptionItemMetaForm.namePlaceholder',
   {
     defaultMessage: 'Name your rule exception',
   }
