+++ x-pack/plugins/security_solution/public/detection_engine/rule_exceptions/api/api_client.ts
@@ -1,0 +1,45 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import { KibanaServices } from '../../../common/lib/kibana';
+
+import { DETECTION_ENGINE_RULES_EXCEPTIONS_REFERENCE_URL } from '../../../../common/constants';
+import type { RulesReferencedByExceptionListsSchema } from '../../../../common/detection_engine/schemas/response';
+
+import type { FindRulesReferencedByExceptionsProps } from './api_client_interface';
+
+/**
+ * Fetch info on what exceptions lists are referenced by what rules
+ *
+ * @param lists exception list information needed for making request
+ * @param signal to cancel request
+ *
+ * @throws An error if response is not OK
+ */
+export const findRuleExceptionReferences = async ({
+  lists,
+  signal,
+}: FindRulesReferencedByExceptionsProps): Promise<RulesReferencedByExceptionListsSchema> => {
+  const idsUndefined = lists.some(({ id }) => id === undefined);
+  const query = idsUndefined
+    ? {
+        namespace_types: lists.map(({ namespaceType }) => namespaceType).join(','),
+      }
+    : {
+        ids: lists.map(({ id }) => id).join(','),
+        list_ids: lists.map(({ listId }) => listId).join(','),
+        namespace_types: lists.map(({ namespaceType }) => namespaceType).join(','),
+      };
+  return KibanaServices.get().http.fetch<RulesReferencedByExceptionListsSchema>(
+    DETECTION_ENGINE_RULES_EXCEPTIONS_REFERENCE_URL,
+    {
+      method: 'GET',
+      query,
+      signal,
+    }
+  );
+};
