+++ x-pack/plugins/security_solution/public/detection_engine/rule_exceptions/api/api_client_interface.ts
@@ -1,0 +1,19 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import type { NamespaceType } from '@kbn/securitysolution-io-ts-list-types';
+
+export interface FindRulesReferencedByExceptionsListProp {
+  id?: string;
+  listId?: string;
+  namespaceType: NamespaceType;
+}
+
+export interface FindRulesReferencedByExceptionsProps {
+  lists: FindRulesReferencedByExceptionsListProp[];
+  signal?: AbortSignal;
+}
