+++ x-pack/plugins/security_solution/public/detection_engine/rule_exceptions/components/exception_item_card/meta.tsx
@@ -21,10 +21,10 @@ import {
 import type { ExceptionListItemSchema } from '@kbn/securitysolution-io-ts-list-types';
 import styled from 'styled-components';
 
+import type { ExceptionListRuleReferencesSchema } from '../../../../../server/lib/detection_engine/rule_exceptions/schemas/response/find_rule_exception_references/schema';
 import * as i18n from './translations';
 import { FormattedDate } from '../../../../common/components/formatted_date';
 import { SecurityPageName } from '../../../../../common/constants';
-import type { ExceptionListRuleReferencesSchema } from '../../../../../common/detection_engine/schemas/response';
 import { SecuritySolutionLinkAnchor } from '../../../../common/components/links';
 import { RuleDetailTabs } from '../../../../detections/pages/detection_engine/rules/details';
 import { getRuleDetailsTabUrl } from '../../../../common/components/link_to/redirect_to_detection_engine';
