+++ x-pack/plugins/security_solution/public/detection_engine/rule_exceptions/components/flyout_components/add_to_lists_table/index.tsx
@@ -11,11 +11,11 @@ import { EuiText, EuiSpacer, EuiInMemoryTable, EuiPanel, EuiLoadingContent } fro
 import type { ExceptionListSchema, ListArray } from '@kbn/securitysolution-io-ts-list-types';
 import { ExceptionListTypeEnum } from '@kbn/securitysolution-io-ts-list-types';
 
-import type { FindRulesReferencedByExceptionsListProp } from '../../../../../detections/containers/detection_engine/rules';
 import * as i18n from './translations';
 import { getSharedListsTableColumns } from '../utils';
 import { useFindExceptionListReferences } from '../../../logic/use_find_references';
-import type { ExceptionListRuleReferencesSchema } from '../../../../../../common/detection_engine/schemas/response';
+import type { ExceptionListRuleReferencesSchema } from '../../../../../../server/lib/detection_engine/rule_exceptions/schemas/response/find_rule_exception_references/schema';
+import type { FindRulesReferencedByExceptionsListProp } from '../../../api/api_client_interface';
 
 interface ExceptionsAddToListsComponentProps {
   /**
