+++ x-pack/plugins/security_solution/server/lib/detection_engine/rule_exceptions/schemas/request/create_rule_default_exceptions/schema.ts
