+++ x-pack/plugins/security_solution/server/lib/detection_engine/rule_exceptions/schemas/request/find_rule_exception_references/schema.test.ts
@@ -6,8 +6,8 @@
  */
 
 import { exactCheck, formatErrors, foldLeftRight } from '@kbn/securitysolution-io-ts-utils';
-import { findExceptionReferencesOnRuleSchema } from './find_exception_list_references_schema';
-import type { FindExceptionReferencesOnRuleSchema } from './find_exception_list_references_schema';
+import { findExceptionReferencesOnRuleSchema } from './schema';
+import type { FindExceptionReferencesOnRuleSchema } from './schema';
 
 describe('find_exception_list_references_schema', () => {
   test('validates all fields', () => {
