+++ x-pack/plugins/security_solution/server/lib/detection_engine/rule_exceptions/schemas/request/find_rule_exception_references/schema.ts
