+++ x-pack/plugins/security_solution/server/routes/index.ts
@@ -72,8 +72,6 @@ import { telemetryDetectionRulesPreviewRoute } from '../lib/detection_engine/rou
 import { readAlertsIndexExistsRoute } from '../lib/detection_engine/routes/index/read_alerts_index_exists_route';
 import { getInstalledIntegrationsRoute } from '../lib/detection_engine/routes/fleet/get_installed_integrations/get_installed_integrations_route';
 import { registerResolverRoutes } from '../endpoint/routes/resolver';
-import { findRuleExceptionReferencesRoute } from '../lib/detection_engine/routes/rules/find_rule_exceptions_route';
-import { createRuleExceptionsRoute } from '../lib/detection_engine/routes/rules/create_rule_exceptions_route';
 import {
   createEsIndexRoute,
   createPrebuiltSavedObjectsRoute,
@@ -84,6 +82,7 @@ import {
   getRiskScoreIndexStatusRoute,
   readPrebuiltDevToolContentRoute,
 } from '../lib/risk_score/routes';
+import { registerRuleExceptionsRoutes } from '../lib/detection_engine/rule_exceptions/api/register_routes';
 export const initRoutes = (
   router: SecuritySolutionPluginRouter,
   config: ConfigType,
@@ -118,7 +117,6 @@ export const initRoutes = (
     previewRuleDataClient,
     getStartServices
   );
-  createRuleExceptionsRoute(router);
 
   // Once we no longer have the legacy notifications system/"side car actions" this should be removed.
   legacyCreateLegacyNotificationRoute(router, logger);
@@ -134,13 +132,14 @@ export const initRoutes = (
 
   registerRuleMonitoringRoutes(router);
 
+  registerRuleExceptionsRoutes(router);
+
   getInstalledIntegrationsRoute(router, logger);
 
   createTimelinesRoute(router, config, security);
   patchTimelinesRoute(router, config, security);
   importRulesRoute(router, config, ml);
   exportRulesRoute(router, config, logger);
-  findRuleExceptionReferencesRoute(router);
 
   importTimelinesRoute(router, config, security);
   exportTimelinesRoute(router, config, security);
